Refactor Countdown
