+++ OpenerCreator/Windows/OpenerCreatorWindow.cs
@@ -1,10 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Numerics;
-using Dalamud.Game;
-using Dalamud.Game.ClientState.Objects.Enums;
-using Dalamud.Interface.Textures;
 using Dalamud.Interface.Windowing;
 using Dalamud.Utility;
 using ImGuiNET;
@@ -17,9 +13,6 @@ namespace OpenerCreator.Windows;
 public class OpenerCreatorWindow : Window, IDisposable
 {
     private static readonly Vector2 IconSize = new(32);
-    private static readonly Vector2 CountdownNumberSize = new(240, 320);
-    private readonly ISharedImmediateTexture countdownGo;
-    private readonly ISharedImmediateTexture countdownNumbers;
 
     private readonly Dictionary<JobCategory, bool> currentColor = new()
     {
@@ -35,7 +28,8 @@ public class OpenerCreatorWindow : Window, IDisposable
     private readonly HashSet<int> wrongActions;
     private int? actionDnd;
     private List<uint> actions;
-    private Stopwatch? countdownStart;
+
+    private Countdown countdown = new();
     private List<string> feedback;
     private List<uint> filteredActions;
     private JobCategory jobCategoryFilter = JobCategory.None;
@@ -70,16 +64,6 @@ public class OpenerCreatorWindow : Window, IDisposable
 
         this.startRecording = startRecording;
         this.stopRecording = stopRecording;
-
-        countdownNumbers = OpenerCreator.TextureProvider.GetFromGame("ui/uld/ScreenInfo_CountDown_hr1.tex");
-        var languageCode = OpenerCreator.DataManager.Language switch
-        {
-            ClientLanguage.French => "fr",
-            ClientLanguage.German => "de",
-            ClientLanguage.Japanese => "ja",
-            _ => "en"
-        };
-        countdownGo = OpenerCreator.TextureProvider.GetFromGame($"ui/icon/121000/{languageCode}/121841_hr1.tex");
     }
 
     public void Dispose()
@@ -98,7 +82,7 @@ public class OpenerCreatorWindow : Window, IDisposable
         DrawSettingsTab();
         ImGui.EndTabBar();
 
-        DrawCountdown();
+        countdown.DrawCountdown();
     }
 
     private void DrawActionsGui()
@@ -324,7 +308,7 @@ public class OpenerCreatorWindow : Window, IDisposable
             feedback.Clear();
             wrongActions.Clear();
             recording = true;
-            countdownStart = Stopwatch.StartNew();
+            countdown.StartCountdown();
             startRecording(OpenerCreator.Config.CountdownTime, AddFeedback, WrongAction);
         }
 
@@ -332,7 +316,7 @@ public class OpenerCreatorWindow : Window, IDisposable
         if (ImGui.Button("Stop Recording"))
         {
             recording = false;
-            countdownStart = null;
+            countdown.StopCountdown();
             stopRecording();
         }
 
@@ -374,66 +358,6 @@ public class OpenerCreatorWindow : Window, IDisposable
 
         ImGui.EndChild();
         ImGui.EndTabItem();
-    }
-
-    private void DrawCountdown()
-    {
-        if (OpenerCreator.Config.IsCountdownEnabled == false || countdownStart == null ||
-            OpenerCreator.ClientState.LocalPlayer!.StatusFlags.ToString()
-                         .Contains(StatusFlags.InCombat.ToString()))
-            return;
-
-        var drawlist = ImGui.GetForegroundDrawList();
-        var timer = OpenerCreator.Config.CountdownTime - (countdownStart.ElapsedMilliseconds / 1000.0f);
-        var ceil = (float)Math.Ceiling(timer);
-        const float uspacing = 1.0f / 6.0f;
-
-        ceil = timer switch
-        {
-            <= 0 => 0,
-            > 5 => (int)Math.Ceiling(timer / 5.0) * 5.0f,
-            _ => ceil
-        };
-
-        var anim = 1.0f - Math.Clamp(ceil - timer - 0.5f, 0.0f, 1.0f);
-        var color = 0x00FFFFFF + ((uint)(anim * 255) << 24);
-
-        if (timer < -2)
-        {
-            countdownStart = null;
-            return;
-        }
-
-        var center = ImGui.GetMainViewport().GetCenter();
-        switch (timer)
-        {
-            case <= 0:
-                drawlist.AddImage(countdownGo.GetWrapOrEmpty().ImGuiHandle,
-                                  center - (countdownGo.GetWrapOrEmpty().Size / 2),
-                                  center + (countdownGo.GetWrapOrEmpty().Size / 2), Vector2.Zero, Vector2.One, color);
-                break;
-            case <= 5:
-                drawlist.AddImage(countdownNumbers.GetWrapOrEmpty().ImGuiHandle, center - (CountdownNumberSize / 2),
-                                  center + (CountdownNumberSize / 2), new Vector2(ceil * uspacing, 0.0f),
-                                  new Vector2((ceil * uspacing) + uspacing, 1.0f), color);
-                break;
-            default:
-            {
-                var dig1 = (int)Math.Floor(ceil / 10.0f);
-                var dig2 = ceil % 10;
-                drawlist.AddImage(countdownNumbers.GetWrapOrEmpty().ImGuiHandle,
-                                  center - CountdownNumberSize with { Y = CountdownNumberSize.Y / 2 },
-                                  center + new Vector2(0.0f, CountdownNumberSize.Y / 2),
-                                  new Vector2(dig1 * uspacing, 0.0f), new Vector2((dig1 * uspacing) + uspacing, 1.0f),
-                                  color);
-                drawlist.AddImage(countdownNumbers.GetWrapOrEmpty().ImGuiHandle,
-                                  center - new Vector2(0.0f, CountdownNumberSize.Y / 2),
-                                  center + CountdownNumberSize with { Y = CountdownNumberSize.Y / 2 },
-                                  new Vector2(dig2 * uspacing, 0.0f), new Vector2((dig2 * uspacing) + uspacing, 1.0f),
-                                  color);
-                break;
-            }
-        }
     }
 
     private void DrawClear()
@@ -503,7 +427,7 @@ public class OpenerCreatorWindow : Window, IDisposable
 
     public void AddFeedback(Feedback f)
     {
-        countdownStart = null;
+        countdown.StopCountdown();
         recording = false;
         feedback = f.GetMessages();
     }
